Fix: card_log write bugs Add Contact pages TODO: fix onchange bug

+++ access_control_system/models/acs_card.py
@@ -22,14 +22,17 @@ class AcsCard(models.Model):
     uid = fields.Char(string='卡片號碼', required=True)
     pin = fields.Char(string='卡片密碼')
 
-    person_ref = fields.Reference( selection=[('res.partner', '客戶') , ('hr.employee', '員工'),], string='用戶')
+    owner_ref = fields.Reference( selection=[('res.partner', '客戶') , ('hr.employee', '員工'),], string='用戶', default=False)
+
+    partner_id = fields.Many2one('res.partner','客戶', default=False)
+    employee_id = fields.Many2one('hr.employee','員工', default=False)
 
     user_code = fields.Char(string='I D',compute='_get_user_code')
     user_name = fields.Char(string='名稱',compute='_get_owner_name')
     user_phone = fields.Char(string='電話',compute='_get_owner_phone')
     user_role = fields.Char(string='身份',compute='_get_owner_role')
 
-    #員工廠商授權進入的門禁群組 改Many2many
+    #卡片被授權進入的門禁群組
     devicegroup_ids = fields.Many2many(
         string='授權門禁群組',
         comodel_name='acs.devicegroup',
@@ -38,86 +41,95 @@ class AcsCard(models.Model):
         column2='card_id',
     )
 
-    partner_id = fields.Many2one('res.partner','客戶')
-    
-    #客戶租用櫃位清單
-    contract_ids = fields.One2many('acs.contract', 'partner_id', string="合約清單")
+    #卡片被授權的合約清單
+    contract_ids = fields.Many2many(
+        string='合約清單',
+        comodel_name='acs.contract',
+        relation='acs_contract_acs_card_rel',
+        column1='contract_id',
+        column2='card_id',
+    )
+    # @api.constrains('devicegroup_ids')
+    # def check_devicegroup_ids(self):
+    #     for record in self:
+    #         if record.user_role == '客戶':
+    #             #客戶卡片不能使用門禁群組
+    #             for dg in record.devicegroup_ids:
+    #                 dg.unlink()
+    #             raise ValidationError('客戶卡片無法更改門禁群組授權')
+    #         else:
+    #             #非客戶卡片不能使用合約列表
+    #             for c in record.contract_ids:
+    #                 c.unlink()
 
-    @api.constrains('devicegroup_ids')
-    def check_devicegroup_ids(self):
-        for record in self:
-            if record.user_role == '客戶':
-                #客戶卡片不能使用門禁群組
-                for dg in record.devicegroup_ids:
-                    dg.unlink()
-                raise ValidationError('客戶卡片無法更改門禁群組授權')
-            else:
-                #非客戶卡片不能使用合約列表
-                for c in record.contract_ids:
-                    c.unlink()
-
-    @api.constrains('contract_ids')
-    def check_contract_ids(self):
-        for record in self:
-            if record.user_role == '客戶':
-                #客戶卡片不能使用門禁群組
-                for dg in record.devicegroup_ids:
-                    dg.unlink()
-            else:
-                #非客戶卡片不能使用合約列表
-                for c in record.contract_ids:
-                    c.unlink()
-                raise ValidationError('只有客戶卡片才能更改合約列表')
+    # @api.constrains('contract_ids')
+    # def check_contract_ids(self):
+    #     for record in self:
+    #         if record.user_role == '客戶':
+    #             #客戶卡片不能使用門禁群組
+    #             for dg in record.devicegroup_ids:
+    #                 dg.unlink()
+    #         else:
+    #             #非客戶卡片不能使用合約列表
+    #             for c in record.contract_ids:
+    #                 c.unlink()
+    #             raise ValidationError('只有客戶卡片才能更改合約列表')
 
 #for compute fields
 #變更用戶欄位時顯示用戶類別
-    @api.onchange('person_ref')
+    @api.onchange('owner_ref')
     def _get_owner_role(self):
         for record in self:
-            _logger.warning( '_get_owner_role! %s' %( record.person_ref ) )
-            if record.person_ref:
-                if 'employee' in record.person_ref:
-                    _logger.warning( 'employee! %s' %( record.person_ref.employee ) )
-                    if (record.person_ref.employee == False):
-                        if hasattr(record.person_ref,'supplier_rank') and record.person_ref.supplier_rank > 0 :
-                            _logger.warning( 'supplier_rank! %s' %( record.person_ref.supplier_rank ) )
-                            record.partner_id = False
+            _logger.warning( '_get_owner_role in: %s' %( record.owner_ref ) )
+            if record.owner_ref:
+                if 'employee' in record.owner_ref:
+                    _logger.warning( 'employee! %s' %( record.owner_ref.employee ) )
+                    if (record.owner_ref.employee == False):
+                        #update partner_id
+                        #record.partner_id = self.env['res.partner'].sudo().search([['id','=',record.owner_ref.id ] ])
+                        record.partner_id = record.owner_ref.id
+                        if hasattr(record.owner_ref,'supplier_rank') and record.owner_ref.supplier_rank > 0 :
+                            _logger.warning( 'supplier_rank! %s' %( record.owner_ref.supplier_rank ) )
                             record.user_role = '廠商'
                         else:
-                            _logger.warning( 'Not supplier! %s' %( record.person_ref.supplier_rank ) )
-                            record.partner_id = record.person_ref
+                            _logger.warning( 'Not supplier! %s' %( record.owner_ref.supplier_rank ) )
                             record.user_role = '客戶'
                 else:
-                    record.partner_id = False
+                    #update employee_id
+                    #record.employee_id  = self.env['hr.employee'].sudo().search([['id','=',record.owner_ref.id ] ])
+                    record.employee_id = record.owner_ref.id
                     record.user_role = '員工'
             else:
+                #update partner_id & employee_id
+                record.partner_id = False
+                record.employee_id = False
                 record.user_role = ''
 
-    @api.onchange('person_ref')
+    @api.onchange('owner_ref')
     def _get_user_code(self):
         for record in self:
-            if record.person_ref:
-                if hasattr(record.person_ref,'vat'):
-                    record.user_code = record.person_ref.vat
+            if record.owner_ref:
+                if hasattr(record.owner_ref,'vat'):
+                    record.user_code = record.owner_ref.vat
                 else:
                     record.user_code = ''
             else:
                 record.user_code = ''
 
-    @api.onchange('person_ref')
+    @api.onchange('owner_ref')
     def _get_owner_name(self):
         for record in self:
-            if record.person_ref:
-                record.user_name = record.person_ref.name
+            if record.owner_ref:
+                record.user_name = record.owner_ref.name
             else:
                 record.user_name = ''
 
-    @api.onchange('person_ref')
+    @api.onchange('owner_ref')
     def _get_owner_phone(self):
         for record in self:
-            if record.person_ref:
-                if hasattr(record.person_ref,'phone'):
-                    record.user_phone = record.person_ref.phone
+            if record.owner_ref:
+                if hasattr(record.owner_ref,'phone'):
+                    record.user_phone = record.owner_ref.phone
                 else:
                     record.user_phone = ''
             else:
@@ -125,19 +137,19 @@ class AcsCard(models.Model):
 #card ORM methods
     @api.model
     def create(self, vals):
-        #write_card_log(self ,vals)
+        write_card_log(self ,vals)
         result = super(AcsCard, self).create(vals)
         return result
     
     def write(self,vals):
-        #write_card_log(self ,vals)
+        write_card_log(self ,vals)
         result = super(AcsCard, self).write(vals)
         return result
 
     def unlink(self):
-        #write_card_log(self ,{})
-        result = super(AcsCard, self).unlink()
-        return result
-
-
-
+        if self.confirmUnlink:
+            write_card_log(self ,{})
+            result = super(AcsCard, self).unlink()
+            return result
+        else:
+            raise ValidationError("禁止未確認的刪除")
